+++ spec/requests/as_content_owner/video_spec.rb
@@ -232,6 +232,21 @@ describe Yt::Video, :partner do
           expect(video.impressions(to: date).keys.max).to eq date.to_date
         end
       end
+
+      specify 'viewer percentages by gender and age range can be retrieved' do
+        expect(video.viewer_percentages[:female]['18-24']).to be_a Float
+        expect(video.viewer_percentages[:female]['25-34']).to be_a Float
+        expect(video.viewer_percentages[:female]['35-44']).to be_a Float
+        expect(video.viewer_percentages[:female]['45-54']).to be_a Float
+        expect(video.viewer_percentages[:female]['55-64']).to be_a Float
+        expect(video.viewer_percentages[:female]['65-']).to be_a Float
+        expect(video.viewer_percentages[:male]['18-24']).to be_a Float
+        expect(video.viewer_percentages[:male]['25-34']).to be_a Float
+        expect(video.viewer_percentages[:male]['35-44']).to be_a Float
+        expect(video.viewer_percentages[:male]['45-54']).to be_a Float
+        expect(video.viewer_percentages[:male]['55-64']).to be_a Float
+        expect(video.viewer_percentages[:male]['65-']).to be_a Float
+      end
     end
   end
 end